pyproject.toml
[project]
name = "kokrokooo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.0.1",
    "black==24.4.2",
    "flake8==6.0.0",
    "flask==3.0.2",
    "flask-login==0.6.3",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf==1.2.1",
    "gunicorn==21.2.0",
    "ipykernel==6.29.0",
    "ipython==8.15.0",
    "ipywidgets==8.1.0",
    "isort==5.12.0",
    "itsdangerous==2.1.2",
    "marshmallow==3.21.0",
    "matplotlib==3.8.0",
    "mysqlclient==2.1.1",
    "nbconvert==7.5.0",
    "notebook==7.0.0",
    "openpyxl==3.1.2",
    "pandas==2.2.1",
    "pillow==12.0.0",
    "plotly==5.18.0",
    "psycopg2-binary==2.9.10",
    "pyinstaller==5.12.0",
    "pytest==7.4.0",
    "pytest-flask==1.3.0",
    "python-dateutil==2.9.0",
    "python-dotenv==1.0.0",
    "python-magic==0.4.27",
    "pytz==2024.1",
    "requests==2.31.0",
    "seaborn==0.13.2",
    "sqlalchemy==2.0.20",
    "sqlalchemy-utils==0.40.0",
    "werkzeug==3.0.1",
    "wtforms==3.1.2",
    "xlrd==2.0.1",
    "zipfile36==0.1.3",
]
